server/requirements.txt
flask
webview
python-dotenv